Fix import problems
